feat: save new franchise

+++ src/main/java/com/prueba/franquicia/repository/FranchiseRepository.java
@@ -3,5 +3,5 @@ package com.prueba.franquicia.repository;
 import com.prueba.franquicia.models.Franchise;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FranchiseRepository extends CrudRepository<Franchise, Integer> {
+public interface FranchiseRepository extends CrudRepository<Franchise, Long> {
 }
